Mixins: Created NodesMixin to share between snakey and networkgraph series.
